hlc: move ParseHLC / DecimalToHLC to util/hlc
util/hlc needs util/log, but ParseHLC/DecimalToHLC doesn't need to be
in tree. To remove the tree dependency on util/log, move the
mentioned functions.

Release note: None
