+++ pkg/ccl/changefeedccl/cdctest/BUILD.bazel
@@ -17,7 +17,6 @@ go_library(
         "//pkg/jobs",
         "//pkg/jobs/jobspb",
         "//pkg/roachpb",
-        "//pkg/sql/sem/tree",
         "//pkg/testutils/serverutils",
         "//pkg/util/fsm",
         "//pkg/util/hlc",
@@ -45,7 +44,6 @@ go_test(
         "//pkg/security",
         "//pkg/security/securitytest",
         "//pkg/server",
-        "//pkg/sql/sem/tree",
         "//pkg/testutils",
         "//pkg/testutils/serverutils",
         "//pkg/testutils/sqlutils",
