+++ pkg/sql/sem/tree/BUILD.bazel
@@ -54,7 +54,6 @@ go_library(
         "function_name.go",
         "grant.go",
         "hide_constants.go",
-        "hlc.go",
         "import.go",
         "indexed_vars.go",
         "insert.go",
@@ -143,7 +142,6 @@ go_library(
         "//pkg/util/duration",
         "//pkg/util/encoding",
         "//pkg/util/errorutil/unimplemented",
-        "//pkg/util/hlc",
         "//pkg/util/ipaddr",
         "//pkg/util/iterutil",
         "//pkg/util/json",
@@ -180,7 +178,6 @@ go_test(
         "expr_test.go",
         "format_test.go",
         "function_name_test.go",
-        "hlc_test.go",
         "indexed_vars_test.go",
         "interval_test.go",
         "main_test.go",
@@ -223,7 +220,6 @@ go_test(
         "//pkg/testutils",
         "//pkg/testutils/sqlutils",
         "//pkg/util/duration",
-        "//pkg/util/hlc",
         "//pkg/util/json",
         "//pkg/util/leaktest",
         "//pkg/util/log",
