+++ pkg/sql/sqlliveness/slstorage/BUILD.bazel
@@ -18,7 +18,6 @@ go_library(
         "//pkg/settings/cluster",
         "//pkg/sql/catalog/descpb",
         "//pkg/sql/sem/eval",
-        "//pkg/sql/sem/tree",
         "//pkg/sql/sqlliveness",
         "//pkg/util/cache",
         "//pkg/util/encoding",
