+++ pkg/util/hlc/BUILD.bazel
@@ -16,10 +16,13 @@ go_library(
     importpath = "github.com/cockroachdb/cockroach/pkg/util/hlc",
     visibility = ["//visibility:public"],
     deps = [
+        "//pkg/sql/pgwire/pgcode",
+        "//pkg/sql/pgwire/pgerror",
         "//pkg/util/log",
         "//pkg/util/syncutil",
         "//pkg/util/timeutil",
         "//pkg/util/tracing",
+        "@com_github_cockroachdb_apd_v3//:apd",
         "@com_github_cockroachdb_errors//:errors",
         "@com_github_cockroachdb_redact//:redact",
     ],
@@ -35,9 +38,12 @@ go_test(
     embed = [":hlc"],
     deps = [
         "//pkg/cli/exit",
+        "//pkg/util/leaktest",
         "//pkg/util/log",
         "//pkg/util/protoutil",
+        "//pkg/util/randutil",
         "//pkg/util/timeutil",
+        "@com_github_cockroachdb_apd_v3//:apd",
         "@com_github_cockroachdb_errors//:errors",
         "@com_github_stretchr_testify//assert",
         "@com_github_stretchr_testify//require",
